requirements.txt
pandas>=2.1.0
numpy>=1.26.0
numpy-financial>=1.0.0
python-dateutil>=2.9.0
sqlalchemy>=2.0.45
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
requests==2.32.5
APScheduler==3.10.4
selenium
webdriver-manager